Change Index type from Int to Long

+++ src/main/scala/co/topl/daml/driver/package.scala
@@ -2,9 +2,7 @@ package co.topl.daml
 
 package object driver {
 
-  // Bifrost supports heights as `Long`, but for indexing purposes here
-  // we will represent offsets as `Int`
-  type Index = Int
+  type Index = Long
 
   val StartIndex: Index = 0
 }
